backend/requirements.txt
bcrypt==4.3.0
blinker==1.8.2
cachelib==0.13.0
click==8.1.8
dnspython==1.16.0
flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-Login==0.6.3
Flask-PyMongo==2.3.0
flask-session==0.8.0
flask-wtf==1.2.1
importlib-metadata==8.5.0
itsdangerous==2.2.0
jinja2==3.1.6
MarkupSafe==2.1.5
msgspec==0.18.6
pymongo==4.10.0
python-dotenv==1.0.1
werkzeug==3.0.6
wtforms==3.1.2
zipp==3.20.2


# Async task system
celery==5.4.0
redis==5.0.8